refactor: split new-game leader dialog into library parts (#3878)
Decompose the 879-line DLG10001 dialog into state-base, slot-row, and
setup-field mixins so the host file stays under the code-review size
target while preserving behavior and widget-test contracts.

Refs #3878
