+++ SPEC/program/repo-and-packages.md
@@ -19,6 +19,7 @@ Monorepo layout:
 - **`AppEventHandlerScope`** (`app/lib/core/services/app_event_handler_scope.dart`) binds session-scoped bus listeners. Session subscriptions are split into `part` files by event family (observe, civilian work, naval/army, diplomacy, debug), each capped at 500 non-comment lines (`repo.app_event_handler_part_size`).
 - **`DebugCommandSessionHandler`** and **`ObserveModeSessionHandler`** apply ctdev/debug and observe-mode session mutations without importing feature panels.
 - **`colonizethis_app_e2e_support`** (`packages/colonizethis_app_e2e_support`) hosts E2E/widget mirror helpers (`e2e_helpers.dart`, `e2e_test_shared*.dart`, panel expected-line fixtures) previously under `app/integration_test/` and `app/test/e2e_*`. `app/test` retains only barrel/contract tests (≤10 `e2e_*.dart` files); integration scenarios import the support package. Seed-42 demo fixtures remain in `app/lib/test_support/` (production/widgetbook consumers).
+- **`features/shell/new_game_leader_selection_dialog.dart`** — DLG10001 new-game leader picker dialog library: `_NewGameLeaderSelectionDialogStateBase` (shared fields + validation), `_NewGameLeaderSelectionDialogSlots` (slot rows + responsive pickers), `_NewGameLeaderSelectionDialogSetupFields` (seed/advanced-start/infinite/terrain/footer). Public widget: `NewGameLeaderSelectionDialog`.
 - **`features/game/flame/caches/`** — shared Flame asset decode helper (`AssetImageCache`) and map marker icon caches (town, port, fleet, civilian, resource, province label). Public barrel: `caches/caches.dart`.
 - **`features/game/flame/map_area/`** — `GameMapAreaBackground` and other render-only map chrome. Public barrel: `map_area/map_area.dart`.
 - **`features/game/flame/map_state/`** — `GameMapArea` widget stack, selection, province actions, turn-feed hooks, and draft projections. `game_map_area.dart` library parts: `game_map_area_turn_feed_labels.dart` (turn-feed display labels + map-locate helpers), `game_map_area_turn_feed.dart` (turn-feed entry builder), `game_map_area_build_map_stack.dart` (map canvas stack + in-map overlay chrome), `game_map_area_build_overlays.dart` (play-area shell + debug/narrow slots), `game_map_area_build.dart` (controls bar + play-area shell). `game_map_area_state_logic.dart` library parts: `game_map_area_state_logic_work_targets.dart`, `game_map_area_state_logic_draft_projection.dart`, `game_map_area_state_logic_province_actions.dart`. `game_map_area_province_action_states.dart` library parts: `game_map_area_province_action_states_prospect.dart`, `game_map_area_province_action_states_explore.dart`, `game_map_area_province_action_states_build_improvement.dart`. `game_map_area_fleet_draft_projection.dart` library parts: `game_map_area_fleet_draft_projection_scope.dart` (location scope keys), `game_map_area_fleet_draft_projection_tiles.dart` (drawable tile resolution). Public barrel: `map_state/map_state.dart`.
